--- dynamic_portfolio_instructions/.ipynb_checkpoints/pages-checkpoint.py
@@ -1,12 +1,0 @@
-from otree.api import Currency as c, currency_range
-from ._builtin import Page, WaitPage
-from .models import Constants
-
-
-class Page1(Page):
-    pass
-
-
-page_sequence = [
-    Page1
-]
