--- dynamic_portfolio_instructions/.ipynb_checkpoints/tests-checkpoint.py
@@ -1,10 +1,0 @@
-from otree.api import Currency as c, currency_range
-from . import pages
-from ._builtin import Bot
-from .models import Constants
-
-
-class PlayerBot(Bot):
-
-    def play_round(self):
-        pass
