--- dynamic_portfolio_instructions/models.py
@@ -1,31 +1,0 @@
-from otree.api import (
-    models, widgets, BaseConstants, BaseSubsession, BaseGroup, BasePlayer,
-    Currency as c, currency_range
-)
-
-
-author = 'Your name here'
-
-doc = """
-Your app description
-"""
-
-
-class Constants(BaseConstants):
-    name_in_url = 'dynamic_portfolio_instructions'
-    players_per_group = None
-    num_rounds = 1
-
-
-class Subsession(BaseSubsession):
-    pass
-
-
-class Group(BaseGroup):
-    pass
-
-
-class Player(BasePlayer):
-    def treatment(self):
-        treatment = self.participant.vars["treatment"]
-        return treatment
