--- dynamic_portfolio_instructions/templates/dynamic_portfolio_instructions/.ipynb_checkpoints/Page1-checkpoint.html
@@ -1,21 +1,0 @@
-{% extends "global/Page.html" %}
-{% load otree static %}
-
-{% block title %}
-    Part 1 Instructions
-{% endblock %}
-
-{% block content %}
- <iframe src="https://docs.google.com/gview?url=https://github.com/jacopomagnani/DynPort_instructions/raw/master/Instructions_part1.pdf&embedded=true" frameborder="0"></iframe>
-    {% next_button %}
-{% endblock %}
-
-{% block styles %}
-    <style>
-        iframe {
-            width: 100%;
-            height: 85vh;
-            padding-bottom: 15px;
-        }
-    </style>
-{% endblock %}
